+++ src/controllers/deployController/deployFilterData.ts
@@ -1,8 +1,8 @@
-import Config from "../../../config";
+import Config from "../../config";
 
-import catchAsync from "../../../utils/asyncHandler";
-import { DeploymentError } from "../../../utils/errors";
-import { createDeploymentDebug } from "../../../utils/createDebug";
+import catchAsync from "../../utils/asyncHandler";
+import { DeploymentError } from "../../utils/errors";
+import { createDeploymentDebug } from "../../utils/createDebug";
 
 const deployFilterData = catchAsync(async (req, res, next) => {
   const debug = createDeploymentDebug(Config.CLIENT_OPTIONS.debug);
