+++ src/routes/login.ts
@@ -2,7 +2,7 @@ import { Router } from "express";
 
 import verifyGithubCode from "../api/middlewares/verifyGithubCode";
 
-import * as AuthController from "../api/controllers/auth";
+import * as AuthController from "../controllers/auth";
 
 const route = Router();
